move Status component to ui-app
