+++ packages/ui-app/src/index.tsx
@@ -43,6 +43,7 @@ import Params from './Params';
 import Password from './Password';
 import Progress from './Progress';
 import Static from './Static';
+import Status from './Status';
 import Tabs from './Tabs';
 
 type Props = BareProps & {
@@ -80,6 +81,7 @@ export {
   Password,
   Progress,
   Static,
+  Status,
   Tabs
 };
